Restore 'Península' region name in mock data
