+++ cheap_electricity/config.py
@@ -1,0 +1,10 @@
+import os
+from typing import Optional
+from dotenv import load_dotenv
+
+# Load environment variables from .env file
+load_dotenv()
+
+ESIOS_API_TOKEN: Optional[str] = os.getenv("ESIOS_API_TOKEN")
+TELEGRAM_BOT_TOKEN: Optional[str] = os.getenv("TELEGRAM_BOT_TOKEN")
+TELEGRAM_CHAT_ID: Optional[str] = os.getenv("TELEGRAM_CHAT_ID")
