+++ tests/test_price_processing.py
@@ -2,7 +2,7 @@ import datetime
 from unittest.mock import patch
 
 from tests.mock_data import MOCK_PRICES_DATA
-from cheap_electricity.main import process_and_categorize_prices
+from cheap_electricity.price_processing import process_and_categorize_prices
 
 
 class FixedDateTimeGreen(datetime.datetime):
@@ -24,21 +24,21 @@ class FixedDateTimeRed(datetime.datetime):
 
 
 def test_process_and_categorize_prices_green():
-    with patch("cheap_electricity.main.datetime.datetime", FixedDateTimeGreen):
+    with patch("cheap_electricity.price_processing.datetime.datetime", FixedDateTimeGreen):
         price, category = process_and_categorize_prices(MOCK_PRICES_DATA)
     assert price == 90.0
     assert category == "Green"
 
 
 def test_process_and_categorize_prices_yellow():
-    with patch("cheap_electricity.main.datetime.datetime", FixedDateTimeYellow):
+    with patch("cheap_electricity.price_processing.datetime.datetime", FixedDateTimeYellow):
         price, category = process_and_categorize_prices(MOCK_PRICES_DATA)
     assert price == 100.0
     assert category == "Yellow"
 
 
 def test_process_and_categorize_prices_red():
-    with patch("cheap_electricity.main.datetime.datetime", FixedDateTimeRed):
+    with patch("cheap_electricity.price_processing.datetime.datetime", FixedDateTimeRed):
         price, category = process_and_categorize_prices(MOCK_PRICES_DATA)
     assert price == 230.0
     assert category == "Red"
